+++ packages/pandas-gbq/docs/source/changelog.rst
@@ -4,6 +4,8 @@ Changelog
 0.2.1 / 2017-??-??
 ------------------
 
+- :func:`read_gbq` now handles query configuration `query.timeoutMs` and stop waiting. (:issue:`76`)
+
 0.2.0 / 2017-07-24
 ------------------